refactor(web): split platform-identity.ts into jwt + routes + service modules
platform-identity.ts was 816 lines (over the 600-line agent-check limit). Split
by responsibility, preserving behavior and the import surface via re-exports:
- platform-jwt.ts: pure OIDC/JWT verification (verifyPlatformJwt,
  assertPlatformMfaClaims, PlatformMfaClaimError).
- platform-identity-routes.ts: registerPlatformIdentityRoutes + OIDC session/
  redirect helpers + login-error page.
- platform-identity.ts: the PlatformIdentityService class, re-exporting the
  moved symbols so existing imports from './platform-identity' still resolve.
Each file is now well under 600 lines (390 / 152 / 302).
